Refactor `internal_moves` module structure
